refactor(PeoplePickerModal): put ModalOverlay inside PeoplePickerControl previously the component that handled the modal's state & the component that rendered a modal were separate moving the ModalOverlay into PeoplePickerContorl also helps to clarify that the rest of what used to be PeoplePickerModal is actually just a particular layout for the PeoplePicker
